+++ src/main/java/gregtech/blocks/tool/BlockBarsTungstenSteel.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019 Gregorius Techneticies
+ * Copyright (c) 2020 GregTech-6 Team
  *
  * This file is part of GregTech.
  *
@@ -50,7 +50,7 @@ public class BlockBarsTungstenSteel extends BlockBaseBars {
 		LH.add(getUnlocalizedName()+ ".15.name", "Tungstensteel Bars");
 	}
 	
-	@Override public float getExplosionResistance(int aMeta) {return 16;}
+	@Override public float getExplosionResistance(byte aMeta) {return 16;}
 	
 	@Override
 	public boolean canEntityDestroy(IBlockAccess aWorld, int aX, int aY, int aZ, Entity aEntity) {
