+++ src/main/java/gregtech/tileentity/plants/MultiTileEntityResinHoleRubber.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019 Gregorius Techneticies
+ * Copyright (c) 2020 GregTech-6 Team
  *
  * This file is part of GregTech.
  *
@@ -75,6 +75,6 @@ public class MultiTileEntityResinHoleRubber extends MultiTileEntityTreeHole {
 	@Override public ItemStack getResinItem(byte aSide) {return IL.IC2_Resin.get(1, IL.Resin.get(1));}
 	@Override public FluidStack getResinFluid(byte aSide) {return FL.make("fluidrubbertreesap", 250);}
 	@Override public ITexture getTexture2(Block aBlock, int aRenderPass, byte aSide, boolean[] aShouldSideBeRendered) {return aShouldSideBeRendered[aSide] ? aSide != mFacing ? BlockTextureCopied.get(BlocksGT.LogA, SIDE_ANY, 0) : BlockTextureDefault.get(mHasResin?Textures.BlockIcons.LOG_RESIN_RUBBER:Textures.BlockIcons.LOG_HOLE_RUBBER) : null;}
-	@Override public float getExplosionResistance2() {return BlocksGT.LogA.getExplosionResistance(0);}
+	@Override public float getExplosionResistance2() {return BlocksGT.LogA.getExplosionResistance((byte)0);}
 	@Override public String getTileEntityName() {return "gt.multitileentity.tree.rubber.resinhole";}
 }
